api 요청 공통처리.

+++ src/components/MyMenu/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { storageItem, StorageKey } from 'api/sign';
+import Cookies from 'js-cookie';
+import { CookieKey, storageItem, StorageKey } from 'api/sign';
 import { useLogout } from 'hooks/useAuth';
 import './styles.css';
-
 interface IProps {
   isOpen: boolean;
   toggleModal: () => void;
@@ -14,8 +14,12 @@ const MyMenu: React.VFC<IProps> = ({ isOpen, toggleModal }) => {
   const logoutMutate = useLogout();
 
   const onLogout = () => {
-    const { value: userId } = storageItem(StorageKey.UserId);
+    const userId = storageItem(StorageKey.UserId);
+
+    if (!userId) return;
+
     window.localStorage.clear();
+    Cookies.remove(CookieKey.RefreshToken);
     logoutMutate(userId);
     toggleModal();
   };
